Move CoinToss implementation for field elements
